requirements.txt
openai
tiktoken
playwright
pillow
matplotlib
jsonlines
pyyaml
opencv-python
google-search-results
pymilvus
transformers